requirements.txt
matplotlib==3.5.1
numpy==1.22.2
Pillow==9.0.1
streamlit==1.7.0
tensorflow==2.8.0
tensorflow_hub==0.12.0
